requirements.txt
psycopg2==2.8.6
gunicorn==19.8.1
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.4
django-ckeditor==6.0.0
django-environ==0.4.5
django-extensions==3.1.0
django-js-asset==1.2.2
Pillow==8.1.0
python-decouple==3.3
pytz==2020.5
static3==0.7.0
django-debug-toolbar==3.2
Pygments==1.6
requests==2.25.1
sqlparse==0.3.1
urllib3==1.26.3